api/quantum_ml.py
import tensorflow as tf
import numpy as np
from typing import Dict, Any, List, Tuple
from sklearn.ensemble import IsolationForest
from sklearn.preprocessing import StandardScaler
import joblib
import os
import time

class QuantumNeuralProcessor:
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.scaler = None
        self.input_shape = config.get('INPUT_SHAPE', (32, 4))
        self._initialize_models()
        self.training_history = []
        self.detection_threshold = 0.95

    def _initialize_models(self):
        """Initialize quantum-inspired neural networks"""
        self.security_model = self._build_security_model()
        self.spyware_model = self._build_spyware_model()
        self.pattern_model = self._build_pattern_model()
        self.anomaly_detector = IsolationForest(
            n_estimators=100,
            contamination=0.1,
            random_state=42
        )

    def _build_security_model(self) -> tf.keras.Model:
        """Build cybersecurity detection model"""
        model = tf.keras.Sequential([
            tf.keras.layers.Input(shape=self.input_shape),
            tf.keras.layers.Conv1D(64, 3, activation='relu', padding='same'),
            tf.keras.layers.BatchNormalization(),
            tf.keras.layers.MaxPooling1D(2),
            tf.keras.layers.LSTM(128, return_sequences=True),
            tf.keras.layers.Dropout(0.3),
            tf.keras.layers.LSTM(64),
            tf.keras.layers.Dense(32, activation='relu'),
            tf.keras.layers.Dropout(0.2),
            tf.keras.layers.Dense(1, activation='sigmoid')
        ])
        
        model.compile(
            optimizer=tf.keras.optimizers.Adam(learning_rate=0.001),
            loss='binary_crossentropy',
            metrics=['accuracy', tf.keras.metrics.AUC()]
        )
        return model

    def _build_spyware_model(self) -> tf.keras.Model:
        """Build spyware detection model"""
        model = tf.keras.Sequential([
            tf.keras.layers.Input(shape=self.input_shape),
            tf.keras.layers.Conv1D(64, 3, activation='relu', padding='same'),
            tf.keras.layers.MaxPooling1D(2),
            tf.keras.layers.Conv1D(128, 3, activation='relu', padding='same'),
            tf.keras.layers.MaxPooling1D(2),
            tf.keras.layers.Flatten(),
            tf.keras.layers.Dense(128, activation='relu'),
            tf.keras.layers.Dropout(0.3),
            tf.keras.layers.Dense(1, activation='sigmoid')
        ])
        
        model.compile(
            optimizer=tf.keras.optimizers.Adam(learning_rate=0.001),
            loss='binary_crossentropy',
            metrics=['accuracy']
        )
        return model

    def _build_pattern_model(self) -> tf.keras.Model:
        """Build pattern recognition model"""
        model = tf.keras.Sequential([
            tf.keras.layers.Input(shape=self.input_shape),
            tf.keras.layers.Conv1D(64, 3, activation='relu', padding='same'),
            tf.keras.layers.BatchNormalization(),
            tf.keras.layers.MaxPooling1D(2),
            tf.keras.layers.Bidirectional(tf.keras.layers.LSTM(64, return_sequences=True)),
            tf.keras.layers.Dropout(0.3),
            tf.keras.layers.Flatten(),
            tf.keras.layers.Dense(64, activation='relu'),
            tf.keras.layers.Dropout(0.2),
            tf.keras.layers.Dense(16, activation='softmax')
        ])
        
        model.compile(
            optimizer=tf.keras.optimizers.Adam(learning_rate=0.001),
            loss='categorical_crossentropy',
            metrics=['accuracy']
        )
        return model

    def _preprocess_data(self, data: np.ndarray) -> np.ndarray:
        """Preprocess data with proper reshaping and scaling"""
        if len(data.shape) == 2:
            data = data.reshape(-1, self.input_shape[0], self.input_shape[1])
        
        if self.scaler is None:
            self.scaler = StandardScaler()
        
        original_shape = data.shape
        flattened = data.reshape(original_shape[0], -1)
        
        if not hasattr(self.scaler, 'mean_'):
            scaled_data = self.scaler.fit_transform(flattened)
        else:
            scaled_data = self.scaler.transform(flattened)
        
        return scaled_data.reshape(original_shape)

    def predict(self, data: np.ndarray) -> np.ndarray:
        """Predict using the security model"""
        processed_data = self._preprocess_data(data)
        return self.security_model.predict(processed_data, verbose=0)

    def adaptive_training(self, data: np.ndarray, labels: np.ndarray) -> Dict[str, Any]:
        """Perform adaptive training with real-time updates"""
        try:
            processed_data = self._preprocess_data(data)
            security_history = self.security_model.fit(
                processed_data, labels,
                epochs=10,
                batch_size=32,
                validation_split=0.2,
                verbose=1
            )
            
            history = {
                'timestamp': time.time(),
                'accuracy': security_history.history['accuracy'][-1],
                'loss': security_history.history['loss'][-1],
                'auc': security_history.history['auc'][-1],
                'success': True
            }
            self.training_history.append(history)
            
            return history
        except Exception as e:
            print(f"Training error: {str(e)}")
            return {
                'error': str(e),
                'success': False
            }

    def detect_threats(self, data: np.ndarray) -> Dict[str, Any]:
        """Detect security threats and anomalies"""
        try:
            processed_data = self._preprocess_data(data)
            security_pred = self.security_model.predict(processed_data, verbose=0)
            spyware_pred = self.spyware_model.predict(processed_data, verbose=0)
            
            flattened_data = processed_data.reshape(processed_data.shape[0], -1)
            anomaly_pred = self.anomaly_detector.fit_predict(flattened_data)
            
            combined_threat_score = (
                0.4 * security_pred +
                0.4 * spyware_pred +
                0.2 * (anomaly_pred == -1).astype(float)
            )
            
            return {
                'threat_detected': bool(combined_threat_score.mean() > self.detection_threshold),
                'threat_score': float(combined_threat_score.mean()),
                'security_confidence': float(security_pred.mean()),
                'spyware_confidence': float(spyware_pred.mean()),
                'is_anomaly': bool(anomaly_pred.mean() == -1),
                'success': True
            }
        except Exception as e:
            print(f"Threat detection error: {str(e)}")
            return {
                'error': str(e),
                'success': False
            }

    def recognize_patterns(self, data: np.ndarray) -> Dict[str, Any]:
        """Recognize complex patterns in data"""
        try:
            processed_data = self._preprocess_data(data)
            pattern_pred = self.pattern_model.predict(processed_data, verbose=0)
            
            dominant_patterns = np.argsort(pattern_pred.mean(axis=0))[-3:]
            pattern_entropy = float(-np.sum(pattern_pred * np.log2(pattern_pred + 1e-10)))
            
            return {
                'dominant_patterns': dominant_patterns.tolist(),
                'pattern_confidences': pattern_pred.mean(axis=0).tolist(),
                'pattern_entropy': pattern_entropy,
                'success': True
            }
        except Exception as e:
            print(f"Pattern recognition error: {str(e)}")
            return {
                'error': str(e),
                'success': False
            }

    def save_models(self, path: str):
        """Save all models and scalers"""
        os.makedirs(path, exist_ok=True)
        self.security_model.save(os.path.join(path, 'security_model'))
        self.spyware_model.save(os.path.join(path, 'spyware_model'))
        self.pattern_model.save(os.path.join(path, 'pattern_model'))
        joblib.dump(self.anomaly_detector, os.path.join(path, 'anomaly_detector.pkl'))
        joblib.dump(self.scaler, os.path.join(path, 'scaler.pkl'))

    def load_models(self, path: str):
        """Load all models and scalers"""
        self.security_model = tf.keras.models.load_model(os.path.join(path, 'security_model'))
        self.spyware_model = tf.keras.models.load_model(os.path.join(path, 'spyware_model'))
        self.pattern_model = tf.keras.models.load_model(os.path.join(path, 'pattern_model'))
        self.anomaly_detector = joblib.load(os.path.join(path, 'anomaly_detector.pkl'))
        self.scaler = joblib.load(os.path.join(path, 'scaler.pkl'))