api/security.py
import numpy as np
from cryptography.hazmat.primitives import hashes
from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2HMAC
from cryptography.hazmat.primitives.ciphers.aead import AESGCM
import os
from typing import Dict, Any, Tuple, List
import hashlib
import hmac
import base64
import jwt
from datetime import datetime, timedelta
from .Config import Config

class QuantumSecurityProtocol:
    def __init__(self, security_level: str = 'high'):
        self.config = Config()
        self.security_level = security_level
        self.quantum_noise_threshold = 0.01
        self.entropy_pool = []
        self._initialize_quantum_security()

    def _initialize_quantum_security(self):
        """Initialize quantum-inspired security components"""
        self.quantum_key = self._generate_quantum_key()
        self.entropy_bits = self._generate_quantum_entropy(1024)
        self.security_layers = self._setup_security_layers()
        self.aesgcm = AESGCM(self.quantum_key)

    def _generate_quantum_key(self, length: int = 32) -> bytes:
        """Generate quantum-inspired encryption key"""
        quantum_random = np.random.randn(length * 8)
        quantum_bits = (quantum_random > self.quantum_noise_threshold).astype(int)
        key_bytes = bytes(int(''.join(map(str, quantum_bits[i:i+8])), 2) 
                         for i in range(0, len(quantum_bits), 8))
        return key_bytes[:length]  # Ensure exact length

    def _generate_quantum_entropy(self, bits: int) -> List[int]:
        """Generate quantum entropy bits"""
        psi = np.random.uniform(0, 2*np.pi, bits)
        quantum_states = np.exp(1j * psi)
        measured_bits = (np.abs(quantum_states) > 1/np.sqrt(2)).astype(int)
        return measured_bits.tolist()

    def _setup_security_layers(self) -> List[Dict[str, Any]]:
        """Setup multiple security layers"""
        return [
            {
                'name': 'quantum_encryption',
                'key': self._generate_quantum_key(),
                'algorithm': 'AES-256-GCM'
            },
            {
                'name': 'quantum_authentication',
                'key': self._generate_quantum_key(),
                'algorithm': 'HMAC-SHA3-512'
            }
        ]

    def encrypt_multi_layer(self, data: bytes) -> Tuple[bytes, Dict[str, bytes]]:
        """Apply multi-layer quantum-inspired encryption"""
        nonce = os.urandom(12)
        metadata = {'nonce': nonce}
        
        noisy_data = self._add_quantum_noise(data)
        encrypted_data = self.aesgcm.encrypt(nonce, noisy_data, None)
        
        auth_tag = self.generate_quantum_signature(encrypted_data)
        metadata['auth_tag'] = auth_tag
        
        return encrypted_data, metadata

    def decrypt_multi_layer(self, encrypted_data: bytes, metadata: Dict[str, bytes]) -> bytes:
        """Decrypt multi-layer quantum-encrypted data"""
        nonce = metadata['nonce']
        auth_tag = metadata['auth_tag']
        
        expected_tag = self.generate_quantum_signature(encrypted_data)
        if not hmac.compare_digest(auth_tag, expected_tag):
            raise ValueError("Authentication failed")
        
        decrypted_data = self.aesgcm.decrypt(nonce, encrypted_data, None)
        original_data = self._remove_quantum_noise(decrypted_data)
        
        return original_data

    def _add_quantum_noise(self, data: bytes) -> bytes:
        """Add quantum-inspired noise for additional security"""
        noise = self._generate_quantum_key(len(data))
        return bytes(d ^ n for d, n in zip(data, noise))

    def _remove_quantum_noise(self, data: bytes) -> bytes:
        """Remove quantum noise from data"""
        noise = self._generate_quantum_key(len(data))
        return bytes(d ^ n for d, n in zip(data, noise))

    def generate_quantum_signature(self, data: bytes) -> bytes:
        """Generate quantum-inspired digital signature"""
        quantum_entropy = bytes(self._generate_quantum_entropy(256))
        signatures = []
        
        for layer in self.security_layers:
            h = hmac.new(layer['key'], data + quantum_entropy, hashlib.sha3_512)
            signatures.append(h.digest())
        
        return b''.join(signatures)

    def generate_token(self, payload: Dict[str, Any]) -> str:
        """Generate a JWT token with quantum-enhanced security"""
        payload['exp'] = datetime.utcnow() + timedelta(hours=self.config['TOKEN_EXPIRE_HOURS'])
        quantum_entropy = self._generate_quantum_entropy(256)
        payload['quantum_entropy'] = base64.b64encode(bytes(quantum_entropy[:16])).decode('utf-8')  # Add entropy
        return jwt.encode(payload, self.config['JWT_SECRET_KEY'], algorithm='HS256')

    def verify_token(self, token: str) -> Tuple[bool, Dict[str, Any]]:
        """Verify a JWT token"""
        try:
            payload = jwt.decode(token, self.config['JWT_SECRET_KEY'], algorithms=['HS256'])
            return True, payload
        except jwt.InvalidTokenError:
            return False, {}