requirements.txt
Flask==2.3.3
flask-cors==5.0.1
numpy==1.26.4
scipy==1.12.0
cryptography==42.0.5
pyjwt==2.8.0
scikit-learn==1.6.1
joblib==1.3.2
kubernetes==29.0.0
aiohttp==3.11.13
Werkzeug